Move `dump_keyring.py` to remove dep `chia.util` => `chia.cmds`
